création des heat-map

+++ scripts/use_case_maxRecordsUser.sh
@@ -1,31 +1,37 @@
 numberOfUserPerIteration=$1
 restartProcess=$2
+END=$3
 
 PATH_FILES=../use-case_NbRecords
 
-rm $PATH_FILES/result_MaxRecords_usecase.csv
+if [ "$restartProcess" = "True" ]
+then
 rm $PATH_FILES/matches.csv
-#rm $PATH_FILES/usersRead.csv
-rm -r $PATH_FILES/usersToRead/
+rm -r $PATH_FILES/test
+rm -r $PATH_FILES/train
+> $PATH_FILES/usersRead.csv
+> $PATH_FILES/result_%MaxRecords_usecase.csv
+fi
 
 i=1
-END=10
 while [ $i -le $END ];
 do
 currentNumberOfUsersRead=$(($i * $numberOfUserPerIteration))
-echo $i
-echo $currentNumberOfUsersRead
+echo
+echo "============NEW ITERATION============"
+echo "Current iteration : $i"
+echo "Current number of users read : $currentNumberOfUsersRead"
+
 #Clean the directories
 rm -r $PATH_FILES/result_ap-attack
-rm -r $PATH_FILES/test
-rm -r $PATH_FILES/train
+rm -r $PATH_FILES/usersToRead/
 #rm $PATH_FILES/final_results.csv
 
 #Create the directory
 mkdir $PATH_FILES/result_ap-attack
 
 #Extract Traces of users with biggest data
-python3 $PATH_FILES/extractUsersWithLargeTraces.py $PATH_FILES/nbRecordsPerUser.csv $PATH_FILES/usersToRead $restartProcess $PATH_FILES/usersRead.csv $currentNumberOfUsersRead
+python3 $PATH_FILES/extractUsersWithLargeTraces.py $PATH_FILES/nbRecordsPerUser.csv $PATH_FILES/usersToRead $numberOfUserPerIteration
 
 #Run split train test
 python3 ../lib_erwan/splitFullTracesInTrainAndTest.py $PATH_FILES/usersToRead/ $PATH_FILES/train $PATH_FILES/test
@@ -38,7 +44,7 @@ OUTPUT_FILE=$(find ../use-case_NbRecords/result_ap-attack/ -type f -iname "run*"
 jq -r '(.report.artifacts[] | select(.name=="MatMatchingKSetsnonObf/matches") | .value) | to_entries | map([.key, .value])[] | @csv' $OUTPUT_FILE | sed 's/"//g' > $PATH_FILES/matches.csv
 
 #Write result to csv
-python3 ../lib_erwan/write_percentage_csv.py $currentNumberOfUsersRead $PATH_FILES/matches.csv
+python3 ../lib_erwan/write_percentage_csv.py $PATH_FILES/matches.csv
 i=$(($i+1))
 done
